requirements.txt
# Reddit API
praw==7.7.1

# Database
supabase==2.3.4

# AI
anthropic==0.18.1

# Web Scraping
firecrawl-py==0.0.16
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
cryptography==42.0.5

# API Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# RSS Parsing (for F5Bot integration)
feedparser==6.0.11

# Rate Limiting
ratelimit==2.2.1

# Async support
aiohttp==3.9.3

# Logging
structlog==24.1.0
